backend/requirements.txt
fastapi
uvicorn
PyPDF2
python-dotenv
langchain
langchain-community
faiss-cpu
pydantic
